+++ lib/sqflite_live.dart
@@ -1,7 +1,7 @@
 /// Support for doing something awesome.
 ///
 /// More dart docs go here.
-library sqflite_live;
+library;
 
 export './src/database_extension.dart';
-export 'package:logger/logger.dart' show Level;
+export './src/host/logger/log_me.dart' show LogLevel;